Rename dynamo
