+++ super-mode-calculator/src/lib/dynamo.ts
